method inlining
