+++ .editorconfig
@@ -1,3 +1,4 @@
 [*.cs]
 indent_size = 4
 indent_style = space
+trim_trailing_whitespace = true
