+++ PooledStream/PooledMemoryStream.netstd11.cs
@@ -1,6 +1,5 @@
 namespace PooledStream
 {
-#if NETSTANDARD1_1
     using System;
     using System.IO;
     using System.Buffers;
@@ -25,30 +24,6 @@ namespace PooledStream
             }
         }
 
-        public override long Seek(long offset, SeekOrigin origin)
-        {
-            long oldValue = _Position;
-            switch ((int)origin)
-            {
-                case (int)SeekOrigin.Begin:
-                    _Position = offset;
-                    break;
-                case (int)SeekOrigin.End:
-                    _Position = _Length - offset;
-                    break;
-                case (int)SeekOrigin.Current:
-                    _Position += offset;
-                    break;
-                default:
-                    throw new InvalidOperationException("unknown SeekOrigin");
-            }
-            if (_Position < 0 || _Position > _Length)
-            {
-                _Position = oldValue;
-                throw new IndexOutOfRangeException();
-            }
-            return _Position;
-        }
         /// <summary>write data to stream</summary>
         /// <remarks>if stream data length is over int.MaxValue, this method throws IndexOutOfRangeException</remarks>
         public override void Write(byte[] buffer, int offset, int count)
@@ -57,7 +32,7 @@ namespace PooledStream
             {
                 throw new InvalidOperationException("stream is readonly");
             }
-            long endOffset = _Position + count;
+            int endOffset = _Position + count;
             if (endOffset > _currentbuffer.Length)
             {
                 ReallocateBuffer((int)(endOffset) * 2);
@@ -72,5 +47,4 @@ namespace PooledStream
         }
 
     }
-#endif
 }
