+++ PooledStream/PooledStream.csproj
@@ -14,10 +14,17 @@
     <PackageProjectUrl>https://github.com/itn3000/PooledStream/</PackageProjectUrl>
     <PackageLicenseUrl>https://opensource.org/licenses/MIT</PackageLicenseUrl>
     <RepositoryUrl>https://github.com/itn3000/PooledStream/</RepositoryUrl>
+    <EnableDefaultCompileItems>false</EnableDefaultCompileItems>
   </PropertyGroup>
   <ItemGroup>
+    <Compile Include="**/*.cs" Exclude="obj/**/*.cs"/>
+    <Compile Include="obj/$(Configuration)/$(TargetFramework)/*.cs"/>
+  </ItemGroup>
+  <ItemGroup Condition="'$(TargetFramework)' == 'netstandard2.1'">
+    <Compile Remove="PooledMemoryStream.netstd11.cs" />
   </ItemGroup>
   <ItemGroup Condition="'$(TargetFramework)' == 'netstandard1.1'">
     <PackageReference Include="System.Buffers" Version="4.5.0" />
+    <Compile Remove="PooledMemoryStream.netstd21.cs" />
   </ItemGroup>
 </Project>
